#95 #132 #133 Extracted DoubleTapGestureDetector class to an external public util class
